+++ scripts/login.sh
@@ -65,45 +65,14 @@ fi
 wcNamespace="org-$wcOrg"
 certIdHash=$(date +"%s" | shasum -a 256)
 certId=${certIdHash::16}
-ttl="1h"
 
 certName="$wcName-$certId"
 wcBaseUrl=${baseUrl:1}
 
-read -r -d '' payload << EOM
-{
-  "apiVersion": "core.giantswarm.io/v1alpha1",
-  "kind": "CertConfig",
-  "metadata": {
-    "labels": {
-      "cert-operator.giantswarm.io/version": "2.0.1",
-      "giantswarm.io/certificate": "$certId",
-      "giantswarm.io/cluster": "$wcName",
-      "giantswarm.io/managed-by": "cluster-operator",
-      "giantswarm.io/organization": "$wcOrg"
-    },
-    "name":"$certName",
-    "namespace":"$wcNamespace"
-  },
-  "spec": {
-    "cert": {
-      "allowBareDomains": true,
-      "clusterComponent": "$certId",
-      "clusterID": "$wcName",
-      "commonName": "$certId.$wcName.k$wcBaseUrl",
-      "disableRegeneration": false,
-      "organizations": ["system:masters"],
-      "ttl":"$ttl"
-    },
-    "versionBundle": {
-      "version":"2.0.1"
-    }
-  }
-}
-EOM
+payload="{\"apiVersion\":\"core.giantswarm.io/v1alpha1\",\"kind\":\"CertConfig\",\"metadata\":{\"labels\":{\"cert-operator.giantswarm.io/version\":\"2.0.1\",\"giantswarm.io/certificate\":\"$certId\",\"giantswarm.io/cluster\":\"$wcName\",\"giantswarm.io/managed-by\":\"cluster-operator\",\"giantswarm.io/organization\":\"$wcOrg\"},\"name\":\"$certName\",\"namespace\":\"$wcNamespace\"},\"spec\":{\"cert\":{\"allowBareDomains\":true,\"clusterComponent\":\"$certId\",\"clusterID\":\"$wcName\",\"commonName\":\"$certId.$wcName.k$wcBaseUrl\",\"disableRegeneration\":false,\"organizations\":[\"system:masters\"],\"ttl\":\"$ttl\"},\"versionBundle\":{\"version\":\"2.0.1\"}}}"
 
 certConfigResponse=$(curl -X POST "$apiUrl/apis/core.giantswarm.io/v1alpha1/namespaces/$wcNamespace/certconfigs" \
-  -d "$(echo "$payload" | jq -c)" -s -H "Content-Type: application/json" \
+  -d "$payload" -s -H "Content-Type: application/json" \
   --header "Authorization: Bearer $idToken" \
   --insecure)
 
@@ -118,10 +87,11 @@ fi
 wcCa=""
 wcCrt=""
 wcKey=""
-attempts=5
 
-while [ $attempts -gt 0 ]
+for attempt in {1..5}
 do
+    echo "Retrieving client certificate data, attempt $attempt/5"
+
     sleep 3
     secretAttemptResponse=$(curl -X GET "$apiUrl/api/v1/namespaces/$wcNamespace/secrets/$certName" --header "Authorization: Bearer $idToken" --insecure -s)
 
@@ -129,19 +99,13 @@ do
     wcCrt=$(echo "$secretAttemptResponse" | jq -r .data.crt)
     wcKey=$(echo "$secretAttemptResponse" | jq -r .data.key)
 
-    if [ -z "$wcCa" ] || [ "$wcCa" == 'null' ]
+    if [ -n "$wcCa" ] && [ "$wcCa" != 'null' ]
     then
-        if [ $attempts == 5 ]
-        then
-          echo "Waiting for client certificate data ..."
-        fi
-        attempts=$(( $attempts - 1 ))
-    else
-        attempts=0
+        break
     fi
 done
 
-if [ "$wcCa" == 'null' ]
+if [ -z "$wcCa" ] || [ "$wcCa" == 'null' ]
 then
     echo "Failed to generate certificate for $wcName in organization $wcOrg"
     exit 1